fixed up instruction type lists
